Created /offices/ endpoint
