+++ express/config/databasePool.ts
@@ -1,7 +1,7 @@
 import { Pool } from 'pg';
 
 const pool = new Pool({
-  user: 'express',
+  user: 'express_user',
   host: process.env.DB_HOST || '127.0.0.1',
   database: process.env.DB_NAME || 'our_civic_voice',
   password: process.env.DB_PASSWORD || 'password',
